HA2P-Net-main/requirements.txt
albumentations==1.1.0
cupy_cuda113==10.6.0
docopt==0.6.2
imgaug==0.4.0
layers==0.1.5
matplotlib==3.4.3
mmcv==2.0.1
numpy==1.22.4
opencv_python==4.5.4.60
opencv_python_headless==4.5.4.60
openxlab==0.0.31
pandas==2.0.3
Pillow==10.1.0
PyYAML==6.0.1
Requests==2.31.0
scikit_image==0.18.3
scipy==1.7.1
tensorboardX==2.6.2.2
thop==0.1.1.post2209072238
tifffile==2021.7.2
timm==0.4.12
torchsummaryX==1.3.0
tqdm==4.65.2
